Cardano.Streaming

Documentation

withChainSyncEventStream

Arguments

:: FilePath
Path to the node socket

-> NetworkId

-> [ChainPoint]
The point on the chain to start streaming from

-> (Stream (Of (ChainSyncEvent (BlockInMode CardanoMode))) IO r -> IO b)
The stream consumer

-> IO b

withChainSyncEventStream uses the chain-sync mini-protocol to connect to a locally running node and fetch blocks from the given starting point.

data ChainSyncEvent a

Constructors

RollForward a ChainTip

RollBackward ChainPoint ChainTip

data ChainSyncEventException

Constructors

NoIntersectionFound

mkConnectInfo :: Env -> FilePath -> LocalNodeConnectInfo CardanoMode

Derive LocalNodeConnectInfo from Env.

mkLocalNodeConnectInfo :: NetworkId -> FilePath -> LocalNodeConnectInfo CardanoMode

Stream blocks and ledger states

blocks :: LocalNodeConnectInfo CardanoMode -> ChainPoint -> Stream (Of (ChainSyncEvent (BlockInMode CardanoMode))) IO r

Create stream of ChainSyncEvent (BlockInMode CardanoMode) from a node at socketPath with networkId starting at point.

blocksPipelined :: Word32 -> LocalNodeConnectInfo CardanoMode -> ChainPoint -> Stream (Of (ChainSyncEvent (BlockInMode CardanoMode))) IO r

ledgerStates :: FilePath -> FilePath -> ValidationMode -> Stream (Of LedgerState) IO r

Get a stream of permanent ledger states

ledgerStatesPipelined :: Word32 -> FilePath -> FilePath -> ValidationMode -> Stream (Of LedgerState) IO r

Get a stream of ledger states over a pipelined chain sync

foldLedgerState :: Env -> LedgerStateHistory -> ValidationMode -> Stream (Of (ChainSyncEvent (BlockInMode CardanoMode))) IO r -> Stream (Of LedgerState) IO r

Fold a stream of blocks into a stream of ledger states. This is implemented in a similar way as foldBlocks in cardano-api:Cardano.Api.LedgerState, the difference being that this keeps waiting for more blocks when chainsync server and client are fully synchronized.

foldLedgerStateEvents :: Env -> LedgerStateHistory -> ValidationMode -> Stream (Of (ChainSyncEvent (BlockInMode CardanoMode))) IO r -> Stream (Of (BlockInMode CardanoMode, LedgerStateEvents)) IO r

Like foldLedgerState, but also produces blocks and LedgerEvents.

getEnvAndInitialLedgerStateHistory :: FilePath -> IO (Env, LedgerStateHistory)

ignoreRollbacks :: Monad m => Stream (Of (ChainSyncEvent a)) m r -> Stream (Of a) m r

Ignore rollback events in the chainsync event stream. Useful for monitor which blocks has been seen by the node, regardless whether they are permanent.