Cardano.Streaming

Documentation

withChainSyncEventStream

Arguments

:: FilePath
Path to the node socket

-> NetworkId

-> [ChainPoint]
The point on the chain to start streaming from

-> (Stream (Of (ChainSyncEvent (BlockInMode CardanoMode))) IO r -> IO b)
The stream consumer

-> IO b

withChainSyncEventStream uses the chain-sync mini-protocol to connect to a locally running node and fetch blocks from the given starting point.

data ChainSyncEvent a

Constructors

RollForward a ChainTip

RollBackward ChainPoint ChainTip

data ChainSyncEventException

Constructors

NoIntersectionFound